@@ -1,29 +1,24 @@
 {
-  "name": "frontend",
+  "name": "getchapull",
   "version": "0.1.0",
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "build": "vue-cli-service build"
   },
   "dependencies": {
-    "@undecaf/vue-material-vuelidate": "^0.2.2",
-    "axios": "^0.20.0",
+    "axios": "^1.7.9",
+    "buffer": "^6.0.3",
     "core-js": "^3.6.5",
     "moment": "^2.29.1",
-    "vue": "^2.6.11",
-    "vue-image-crop-upload": "^2.5.0",
-    "vue-router": "^3.4.7"
+    "vue": "^3.5.13",
+    "vue-image-crop-upload": "^3.0.3",
+    "vue-router": "^4.5.0",
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-service": "~4.5.0",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.2.2",
-    "vue-template-compiler": "^2.6.11"
+    "@vue/cli-plugin-babel": "^5.0.8",
+    "@vue/cli-service": "^5.0.8"
   },
   "eslintConfig": {
     "root": true,
@@ -31,11 +26,11 @@
       "node": true
     },
     "extends": [
-      "plugin:vue/essential",
+      "plugin:vue/vue3-essential",
       "eslint:recommended"
     ],
     "parserOptions": {
-      "parser": "babel-eslint"
+      "parser": "@babel/eslint-parser"
     },
     "rules": {}
   },
