@@ -1,50 +1,50 @@
-import Vue from 'vue'
-import App from './App.vue'
-import VueRouter from 'vue-router'
-import Login from './components/Login.vue'
-import SignUp from './components/Signup.vue'
-import NavComponent from './components/NavBarComponent.vue'
-import BaseContent from './components/ProfileBaseComponent.vue'
-import EditProfile from './components/EditProfileComponent.vue'
-import MainPage from './components/MainPage.vue'
+import { createApp } from 'vue';
+import App from './App.vue';
+import { createRouter, createWebHistory } from 'vue-router'; // VueRouter yerine bu import edilir
+import Login from './components/Login.vue';
+import SignUp from './components/Signup.vue';
+import NavComponent from './components/NavBarComponent.vue';
+import BaseContent from './components/ProfileBaseComponent.vue';
+import EditProfile from './components/EditProfileComponent.vue';
+import MainPage from './components/MainPage.vue';
+import { Buffer } from 'buffer';
 
-Vue.config.productionTip = false
-Vue.use(VueRouter)
+window.Buffer = Buffer;
+const BASE_URL = "http://localhost:8080";
 
-// Vue.prototype.$url = "https://metal-review-spring.herokuapp.com";
-Vue.prototype.$url = "http://localhost:8080";
+const routes = [
+  { path: '/login', component: Login },
+  { path: '/signup', component: SignUp },
+  {path: '/', component: NavComponent,
+    children: [
+      {
+        path: '',
+        component: MainPage
+      },
+      {
+        path: ':username',
+        component: BaseContent
+      },
+      {
+        path: ':username/edit-profile',
+        component: EditProfile
+      }
+    ]
+  }
+];
 
-const router = new VueRouter({
-  routes: [
-    { path: '/login', component: Login},
-    { path: '/signup', component: SignUp},
-    { path: '/', component: NavComponent,
-      children: [
-        {
-          path: '',
-          component: MainPage
-        },
-        {
-          path: ':username',
-          component: BaseContent
-        },
-        {
-          path: ':username/edit-profile',
-          component: EditProfile
-        }
-      ]
-    }
-  ],
-  mode: 'history'
-})
+const router = createRouter({
+  history: createWebHistory(),
+  routes
+});
 
 Date.prototype.addDays = function(days) {
-  var date = new Date(this.valueOf());
+  const date = new Date(this.valueOf());
   date.setDate(date.getDate() + days);
   return date;
-}
+};
 
-router.beforeEach( (to, from, next) => {
+router.beforeEach((to, from, next) => {
   const expirationDate = localStorage.getItem('expirationDate');
   const now = new Date();
   if( now >= expirationDate) next( () => {
@@ -54,9 +54,11 @@ router.beforeEach( (to, from, next) => {
     router.push('/login')
   })
   else next()
-})
+});
 
-new Vue({
-  render: h => h(App),
-  router
-}).$mount('#app')
+const app = createApp(App);
+
+app.config.globalProperties.$url = BASE_URL;
+
+app.use(router);
+app.mount('#app');
